+++ packages/sdk/src/sync/closableAsyncIterator.ts
@@ -26,4 +26,40 @@ export class ClosableAsyncIterator<T> {
 
     return new ClosableAsyncIterator(mapped(), async () => await this.close());
   }
+
+  // iterate over the iterator in batches of size `batchSize`.
+  // `exact` defaults to false
+  // if `exact` is set to `false`, upon closing, any leftover elements will be emitted
+  // that aren't enough to fill a batch will be emitted as a final batch of size < `batchSize`.
+  // if `exact` is set to true, then any leftover elements will be discarded.
+  batches(batchSize: number, exact = false): ClosableAsyncIterator<T[]> {
+    const thisIter = this.iter;
+    const batched = async function* () {
+      let batch: T[] = [];
+      for await (const item of thisIter) {
+        batch.push(item);
+        if (batch.length === batchSize) {
+          yield batch;
+          batch = [];
+        }
+      }
+      if (!exact && batch.length > 0) yield batch;
+    };
+
+    return new ClosableAsyncIterator(batched(), async () => await this.close());
+  }
+
+  // execute a function over each item in the iterator without consuming the iterator or modifying the values
+  // NOTE: the reference given to `f` will allow mutation of the underlying value in the iterator.
+  tap(f: (item: T) => void): ClosableAsyncIterator<T> {
+    const thisIter = this.iter;
+    const tapped = async function* () {
+      for await (const item of thisIter) {
+        f(item);
+        yield item;
+      }
+    };
+
+    return new ClosableAsyncIterator(tapped(), async () => await this.close());
+  }
 }