+++ packages/subgraph/src/handler.ts
@@ -1,4 +1,6 @@
 import {
+  InsertNote,
+  InsertNoteCommitments,
   JoinSplitProcessed,
   RefundProcessed,
   SubtreeUpdate,
@@ -9,6 +11,8 @@ import {
   EncryptedNote,
   SubtreeCommit,
   Nullifier,
+  CompressedEncodedNote,
+  TreeInsertion,
 } from "../generated/schema";
 import {
   getTotalLogIndex,
@@ -130,4 +134,37 @@ export function handleSubtreeUpdate(event: SubtreeUpdate): void {
   commit.subtreeIndex = event.params.subtreeIndex;
   commit.idx = idx;
   commit.save();
+}
+
+export function handleInsertNote(event: InsertNote): void {
+  const totalLogIndex = getTotalLogIndex(event);
+  const idx = getTotalEntityIndex(totalLogIndex, 0);
+  const id = toPadded32BArray(idx);
+
+  const compressedNote = new CompressedEncodedNote(id);
+  compressedNote.ownerH1 = event.params.note.ownerH1;
+  compressedNote.ownerH2 = event.params.note.ownerH2;
+  compressedNote.nonce = event.params.note.nonce;
+  compressedNote.encodedAssetAddr = event.params.note.encodedAssetAddr;
+  compressedNote.encodedAssetId = event.params.note.encodedAssetId;
+  compressedNote.value = event.params.note.value;
+  compressedNote.save();
+
+  const insertion = new TreeInsertion(id);
+  insertion.note = id;
+  insertion.idx = idx;
+  insertion.save();
+}
+
+export function handleInsertNoteCommitments(
+  event: InsertNoteCommitments
+): void {
+  const totalLogIndex = getTotalLogIndex(event);
+  const idx = getTotalEntityIndex(totalLogIndex, 0);
+  const id = toPadded32BArray(idx);
+
+  const insertion = new TreeInsertion(id);
+  insertion.idx = idx;
+  insertion.noteCommitments = event.params.commitments;
+  insertion.save();
 }