+++ packages/subtree-updater/src/types.ts
@@ -1,0 +1,15 @@
+import { BaseProof, SubtreeUpdateInputs } from "@nocturne-xyz/sdk";
+
+export interface ProofJobData {
+  subtreeIndex: number;
+  newRoot: bigint;
+  proofInputs: SubtreeUpdateInputs;
+}
+
+export interface SubmissionJobData {
+  proof: BaseProof;
+  newRoot: bigint;
+}
+
+export type SerializedProofJobData = string;
+export type SerializedSubmissionJobData = string;