+++ packages/sdk/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+- add `fromCompressedPoints` method to `StealthAddress` trait
+- add `batches` method to `ClosableAsyncIterator`
 - remove `protocolWhitelistKey` fn after removing selectors from whitelist
 - add optional `throttleMs` field to `IterSyncOpts` and implement it in all sync adapters
 - subgraph fetch functions query via `idx_gte` and `idx_lt` instead of `id_gte` and `id_lt`